refactor(fe): refactor search bar
