+++ apps/frontend/app/(client)/(main)/_components/SearchBar.tsx
@@ -14,53 +14,63 @@ interface Inputs {
 
 interface SearchBarProps {
   className?: string
+  height?: 'sm' | 'md' | 'lg'
+  fontSize?: 'base' | 'lg'
 }
 
-export function SearchBar({ className }: SearchBarProps) {
+const heightClassMap: Record<NonNullable<SearchBarProps['height']>, string> = {
+  sm: '!h-[36px]',
+  md: '!h-[40px]',
+  lg: '!h-[46px]'
+}
+
+const buildUrl = (pathname: string, params: URLSearchParams): Route =>
+  `${pathname}${params.size ? `?${params}` : ''}` as Route
+
+export function SearchBar({
+  className,
+  height = 'sm',
+  fontSize
+}: SearchBarProps) {
   const router = useRouter()
   const pathname = usePathname()
-  const searchParam = useSearchParams()
-  const search = searchParam.get('search')
-  const { register, handleSubmit, setValue } = useForm<Inputs>()
-  const onSubmit = (data: Inputs) => {
-    const newParam = new URLSearchParams()
-    // set all searchParam to newParam
-    searchParam.forEach((value, key) => newParam.set(key, value))
-    // set search data to newParam
-    if (data.search) {
-      newParam.set('search', data.search)
-    } else {
-      newParam.delete('search')
-    }
-    const newParamString = newParam.toString()
-    router.push(
-      `${pathname}${newParamString ? `?${newParamString}` : ''}` as Route,
-      {
-        scroll: false
-      }
-    )
-  }
+  const searchParams = useSearchParams()
+  const search = searchParams.get('search')
+
+  const { register, handleSubmit, reset } = useForm<Inputs>({
+    defaultValues: { search: search ?? '' }
+  })
 
   useEffect(() => {
-    // set search value from searchParam when page is loaded
-    setValue('search', search ?? '')
-  }, [search, setValue])
+    reset({ search: search ?? '' })
+  }, [search, reset])
+
+  const onSubmit = ({ search }: Inputs) => {
+    const newParams = new URLSearchParams(searchParams.toString())
+    search ? newParams.set('search', search) : newParams.delete('search')
+    router.push(buildUrl(pathname, newParams), { scroll: false })
+  }
 
   return (
     <form
-      className={cn('relative', className)}
+      className={cn('relative min-w-[184px] max-w-[580px]', className)}
       onSubmit={handleSubmit(onSubmit)}
     >
       <IoSearch
-        className="text-muted-foreground absolute left-3 top-[10px] h-4 w-4"
+        className="text-muted-foreground absolute left-3 top-1/2 h-4 w-4 -translate-y-1/2"
         color="#C4C4C4"
       />
       <Input
         placeholder="Search"
         sizeVariant="sm"
-        className="rounded-full px-8 placeholder:text-[#8A8A8A]"
+        className={cn(
+          'rounded-full px-8 placeholder:text-[#C4C4C4]',
+          heightClassMap[height],
+          fontSize === 'lg'
+            ? 'text-lg placeholder:text-lg'
+            : 'text-base placeholder:text-base'
+        )}
         {...register('search')}
-        defaultValue={search ?? ''}
       />
     </form>
   )
